+++ src/main/kotlin/org/radarbase/push/integration/fitbit/service/fitbitapi/FitbitRequestGenerator.kt
@@ -1,0 +1,4 @@
+package org.radarbase.push.integration.fitbit.service.fitbitapi
+
+class FitbitRequestGenerator {
+}
